# PERF: prevents eager pluck in post alerter